Удалены лишние отступы, создана папка "data" для файлов, создан файл для функций
