+++ data/companies_liquidated_result.csv
